- Build three variations of the kernel: generic, sse4, avx2 - Dynamically choose the best kernel to load in bootloader - Generate disassembly and bochs symbols for all variations - Add bochs launcher make targets for {fat/iso}{kernel/boot} - Add bochs startup debugger scripts - Align IDTR loads - Add more CPU detection code to bootloader for selecting kernel - Add bios, keyboard, port92 detection for A20 enable/disable - Cleanup copy_or_enter - Add protected mode fastpath for loading below 4GB line - Cleanup farptr code - Speedup paging code in bootloader - Small improvement in utf8 code in bootloader - Add AMD CPU topology detection code - Stop printing message too early in cpu_init - Add CPU constants header for use in assembly (fix hardcoding) - Add nano_time helpers for translating TSC to/from ns - Fix incorrect handling of 64-bit ACPI tables - Fix incorrect handling of multiple IOAPICs - Add __generic_target macro for disabling fancy instructions on a per function basis - Make cpuid code callable from assembly - Initialize AVX enough to work earlier, from entry.S - Add size parameter to mpresent for validating that a range is accessible - Add a compiler barrier in spinlock_unlock - Use double underscore versions of attribute names - Add serial output to e9debug code - Finish implementation of polling serial code for super early logging - Add callout "constructors_ran" for code dependent upon global constructors being complete - Fix crash when trying to output debug string too early - Change annoying misleading symbol name in bootfat.s - Return byte in bioscall ah_if_carry implementation

+++ boot/bioscall.h
@@ -29,7 +29,7 @@ struct bios_regs_t {
         return eflags_out & (1 << 0);
     }
 
-    uint16_t ah_if_carry() const
+    uint8_t ah_if_carry() const
     {
         if (flags_CF())
             return (eax >> 8) & 0xFF;
